move StringArrayType, StringViewArrayBuilder, etc outside of string module
