+++ datafusion/functions/src/lib.rs
@@ -135,6 +135,8 @@ make_stub_package!(unicode, "unicode_expressions");
 #[cfg(any(feature = "datetime_expressions", feature = "unicode_expressions"))]
 pub mod planner;
 
+pub mod strings;
+
 mod utils;
 
 /// Fluent-style API for creating `Expr`s
